pyproject.toml
[project]
name = "linux-whisper"
version = "0.1.0"
description = "Wispr Flow-quality voice dictation, running entirely locally on Linux"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Nathan Ramia" }]

dependencies = [
    # Core (always required)
    "numpy>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# ML backends (install what you need)
audio = ["sounddevice>=0.5"]
vad = ["onnxruntime>=1.19"]
moonshine = ["moonshine>=0.1"]
llm = ["llama-cpp-python>=0.3"]
hotkey = ["evdev>=1.7"]
tray = ["pystray>=0.19", "Pillow>=10.0"]
whisper = ["pywhispercpp>=1.4"]
rocm = ["onnxruntime-rocm>=1.19"]

# Install everything for full functionality
full = [
    "linux-whisper[audio,vad,whisper,moonshine,llm,hotkey,tray]",
]

dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
]

[project.scripts]
linux-whisper = "linux_whisper._entry:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/linux_whisper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that require model loading",
    "integration: marks integration tests requiring full pipeline",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
